set_timeout must update timeout in watchdog_device
the driver wasn't storing the new timeout value in the timeout field of struct watchdog_device,
causing userspace programs to assume the wrong value for timeout was used (watchdog_device.timeout
still contained the default timeout value) when using
   ioctl(fd, WDIOC_SETTIMEOUT, &timeout);
which updates the caller's timeout variable.

This caused the hardware WDT to reset the system if the new set_timeout value is less than the driver's default.